+++ modules/dnn/src/precomp.hpp
@@ -85,6 +85,7 @@
 #include <opencv2/core/opencl/ocl_defs.hpp>
 
 #include <opencv2/core/utils/trace.hpp>
+#include <opencv2/imgproc.hpp>
 #include <opencv2/dnn.hpp>
 #include <opencv2/dnn/all_layers.hpp>
 #include <opencv2/dnn/shape_utils.hpp>
