+++ modules/imgproc/include/opencv2/imgproc.hpp
@@ -4390,6 +4390,145 @@ at most 8 vertices. Stored as std::vector\<cv::Point2f\> or cv::Mat as Mx1 of ty
  */
 CV_EXPORTS_W int rotatedRectangleIntersection( const RotatedRect& rect1, const RotatedRect& rect2, OutputArray intersectingRegion  );
 
+/** @brief Performs non maximum suppression given boxes and corresponding scores.
+
+* @param bboxes a set of bounding boxes to apply NMS.
+* @param scores a set of corresponding confidences.
+* @param score_threshold a threshold used to filter boxes by score.
+* @param nms_threshold a threshold used in non maximum suppression.
+* @param indices the kept indices of bboxes after NMS.
+* @param eta a coefficient in adaptive threshold formula: \f$nms\_threshold_{i+1}=eta\cdot nms\_threshold_i\f$.
+* @param top_k if `>0`, keep at most @p top_k picked indices.
+*/
+CV_EXPORTS void NMSBoxes(const std::vector<Rect>& bboxes, const std::vector<float>& scores,
+                            const float score_threshold, const float nms_threshold,
+                            CV_OUT std::vector<int>& indices,
+                            const float eta = 1.f, const int top_k = 0);
+
+CV_EXPORTS_W void NMSBoxes(const std::vector<Rect2d>& bboxes, const std::vector<float>& scores,
+                            const float score_threshold, const float nms_threshold,
+                            CV_OUT std::vector<int>& indices,
+                            const float eta = 1.f, const int top_k = 0);
+
+CV_EXPORTS_AS(NMSBoxesRotated) void NMSBoxes(const std::vector<RotatedRect>& bboxes, const std::vector<float>& scores,
+                            const float score_threshold, const float nms_threshold,
+                            CV_OUT std::vector<int>& indices,
+                            const float eta = 1.f, const int top_k = 0);
+
+/** @brief Performs batched non maximum suppression on given boxes and corresponding scores across different classes.
+
+* @param bboxes a set of bounding boxes to apply NMS.
+* @param scores a set of corresponding confidences.
+* @param class_ids a set of corresponding class ids. Ids are integer and usually start from 0.
+* @param score_threshold a threshold used to filter boxes by score.
+* @param nms_threshold a threshold used in non maximum suppression.
+* @param indices the kept indices of bboxes after NMS.
+* @param eta a coefficient in adaptive threshold formula: \f$nms\_threshold_{i+1}=eta\cdot nms\_threshold_i\f$.
+* @param top_k if `>0`, keep at most @p top_k picked indices.
+*/
+CV_EXPORTS void NMSBoxesBatched(const std::vector<Rect>& bboxes, const std::vector<float>& scores, const std::vector<int>& class_ids,
+                                const float score_threshold, const float nms_threshold,
+                                CV_OUT std::vector<int>& indices,
+                                const float eta = 1.f, const int top_k = 0);
+
+CV_EXPORTS_W void NMSBoxesBatched(const std::vector<Rect2d>& bboxes, const std::vector<float>& scores, const std::vector<int>& class_ids,
+                                    const float score_threshold, const float nms_threshold,
+                                    CV_OUT std::vector<int>& indices,
+                                    const float eta = 1.f, const int top_k = 0);
+
+/**
+* @brief Enum of Soft NMS methods.
+* @see softNMSBoxes
+*/
+enum class SoftNMSMethod
+{
+    SOFTNMS_LINEAR = 1,
+    SOFTNMS_GAUSSIAN = 2
+};
+
+/** @brief Performs soft non maximum suppression given boxes and corresponding scores.
+* Reference: https://arxiv.org/abs/1704.04503
+* @param bboxes a set of bounding boxes to apply Soft NMS.
+* @param scores a set of corresponding confidences.
+* @param updated_scores a set of corresponding updated confidences.
+* @param score_threshold a threshold used to filter boxes by score.
+* @param nms_threshold a threshold used in non maximum suppression.
+* @param indices the kept indices of bboxes after NMS.
+* @param top_k keep at most @p top_k picked indices.
+* @param sigma parameter of Gaussian weighting.
+* @param method Gaussian or linear.
+* @see SoftNMSMethod
+*/
+CV_EXPORTS_W void softNMSBoxes(const std::vector<Rect>& bboxes,
+                                const std::vector<float>& scores,
+                                CV_OUT std::vector<float>& updated_scores,
+                                const float score_threshold,
+                                const float nms_threshold,
+                                CV_OUT std::vector<int>& indices,
+                                size_t top_k = 0,
+                                const float sigma = 0.5,
+                                SoftNMSMethod method = SoftNMSMethod::SOFTNMS_GAUSSIAN);
+
+namespace detail {
+
+/** @brief Helper function for NMS
+*/
+CV_EXPORTS void NMSGetMaxScoreIndex(const std::vector<float>& scores, const float threshold, const int top_k,
+                      std::vector<std::pair<float, int> >& score_index_vec);
+
+/** @brief Do non-maximum suppression given bboxes and scores.
+
+Inspired by Piotr Dollar's NMS implementation in EdgeBox.
+https://goo.gl/jV3JYS
+   @param bboxes a set of bounding boxes.
+   @param scores a set of corresponding confidences.
+   @param score_threshold a threshold used to filter detection results.
+   @param nms_threshold a threshold used in non maximum suppression.
+   @param eta
+   @param top_k if not > 0, keep at most top_k picked indices.
+   @param indices the kept indices of bboxes after nms.
+   @param computeOverlap
+   @param limit early terminate once the # of picked indices has reached it.
+*/
+template <typename BoxType>
+void NMSBoxesFast(const std::vector<BoxType>& bboxes,
+      const std::vector<float>& scores, const float score_threshold,
+      const float nms_threshold, const float eta, const int top_k,
+      std::vector<int>& indices,
+      float (*computeOverlap)(const BoxType&, const BoxType&),
+      size_t limit = std::numeric_limits<int>::max())
+{
+    CV_Assert(bboxes.size() == scores.size());
+
+    // Get top_k scores (with corresponding indices).
+    std::vector<std::pair<float, int> > score_index_vec;
+    detail::NMSGetMaxScoreIndex(scores, score_threshold, top_k, score_index_vec);
+
+    // Do nms.
+    float adaptive_threshold = nms_threshold;
+    indices.clear();
+    for (size_t i = 0; i < score_index_vec.size(); ++i) {
+        const int idx = score_index_vec[i].second;
+        bool keep = true;
+        for (int k = 0; k < (int)indices.size() && keep; ++k) {
+            const int kept_idx = indices[k];
+            float overlap = computeOverlap(bboxes[idx], bboxes[kept_idx]);
+            keep = overlap <= adaptive_threshold;
+        }
+        if (keep) {
+            indices.push_back(idx);
+            if (indices.size() >= limit) {
+                break;
+            }
+        }
+        if (keep && eta < 1 && adaptive_threshold > 0.5) {
+          adaptive_threshold *= eta;
+        }
+    }
+}
+
+} // detail::
+
 /** @brief Creates a smart pointer to a cv::GeneralizedHoughBallard class and initializes it.
 */
 CV_EXPORTS_W Ptr<GeneralizedHoughBallard> createGeneralizedHoughBallard();
